Command Manager Update + API Addition
+ All commands handled through the command manager.
+ All permissions handled though plugin.yml
+ Added getPoints() method for developers (returns balance)
- Old command manager & commands
*plugin.yml new desc weren't filled.
