--- src/main/java/com/walrusone/skywars/commands/StartCommand.java
@@ -1,51 +1,0 @@
-package com.walrusone.skywars.commands;
-
-import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
-import com.walrusone.skywars.SkyWarsReloaded;
-import com.walrusone.skywars.game.Game;
-import com.walrusone.skywars.game.Game.GameState;
-import com.walrusone.skywars.game.GamePlayer;
-import com.walrusone.skywars.utilities.Messaging;
-
-public class StartCommand implements CommandExecutor {
-	
-	@Override
-    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-		boolean hasPerm = false;
-		if (!(sender instanceof Player)) {
-			hasPerm = true;
-		} else if (sender instanceof Player) {
-			Player player = (Player) sender;
-			if (SkyWarsReloaded.perms.has(player, "swr.start")) {
-				hasPerm = true;
-			}
-		} else {
-			sender.sendMessage(new Messaging.MessageFormatter().format("error.cmd-no-perm"));
-		}
-		if (hasPerm) {
-			if (args.length == 1) {
-				if (sender instanceof Player) {
-					Player player = (Player) sender;
-					GamePlayer gPlayer = SkyWarsReloaded.getPC().getPlayer(player.getUniqueId());
-					if (gPlayer.inGame()) {
-						Game game = gPlayer.getGame();
-						if (game.getState() == GameState.PREGAME) {
-							game.startGame();
-						}
-					} else {
-						sender.sendMessage(new Messaging.MessageFormatter().format("error.not-in-game"));
-					}
-				}
-			} else {
-				sender.sendMessage(ChatColor.RED + "USAGE: /swr start");
-			}
-		} 
-		return true;
-	}
-	
-}
